refactor(auth): package 구조 변경
entity 패키지 안에 auth 패키지를 생성
